+++ python/lsst/ts/gis/csc.py
@@ -3,10 +3,16 @@ __all__ = ["GISCsc"]
 import asyncio
 
 from lsst.ts import salobj, utils
+from pymodbus.datastore import (
+    ModbusSequentialDataBlock,
+    ModbusServerContext,
+    ModbusSlaveContext,
+)
+from pymodbus.server import ModbusTcpServer
 
-from .config import CONFIG_SCHEMA
-from .component import GISComponent
 from . import __version__
+from .component import GISComponent
+from .config import CONFIG_SCHEMA
 
 
 class GISCsc(salobj.ConfigurableCsc):
@@ -15,18 +21,28 @@ class GISCsc(salobj.ConfigurableCsc):
     Parameters
     ----------
     initial_state : `lsst.ts.salobj.State`
-    settings_to_apply : `str`
+        The initial state that the CSC starts in.
+    override : `str`
+        The name of the config file.
     simulation_mode : `int`
+        Is the GIS in simulation mode?
     config_dir : `pathlib.Path`
+        The custom configuration file directory.
 
     Attributes
     ----------
+    mock_server : `pymodbus.server.ModbusTcpServer`
+        The mock modbus server.
+        Just returns static values.
     component : `GISComponent`
-    simulator : `None`
+        Handles the data received from the GIS.
+    telemetry_interval : `float`
+        The time to sleep for telemetry publishing.
     telemetry_task : `asyncio.Future`
+        An asyncio task to handle starting and cancelling the telemetry loop.
     """
 
-    valid_simulation_modes = [0]
+    valid_simulation_modes = [0, 1]
     version = __version__
 
     def __init__(
@@ -45,6 +61,7 @@ class GISCsc(salobj.ConfigurableCsc):
             simulation_mode=simulation_mode,
             config_dir=config_dir,
         )
+        self.mock_server = None
         self.component = GISComponent(self)
         self.simulator = None
         self.telemetry_interval = None
@@ -53,8 +70,12 @@ class GISCsc(salobj.ConfigurableCsc):
     async def telemetry_loop(self):
         """Implement the telemetry feed for the GIS."""
         while True:
-            await self.component.update_status()
-            await asyncio.sleep(self.telemetry_interval)
+            try:
+                await self.component.update_status()
+                await asyncio.sleep(self.telemetry_interval)
+            except Exception:
+                self.log.exception("Telemetry loop failed.")
+                raise
 
     @property
     def connected(self):
@@ -62,21 +83,44 @@ class GISCsc(salobj.ConfigurableCsc):
         return self.component.connected
 
     async def handle_summary_state(self):
+        """Handle summary state transitions.
+
+        When transitioning to disabled or enabled state.
+        * Connect to the client if not connected.
+        * Start the mock server if in simulation mode and not started.
+        * Start the telemetry loop if not started.
+
+        When outside of the disabled or enabled state.
+
+        * disconnect from the client if connected.
+        * Close the mock server if running.
+        * Cancel the telemetry loop.
+        """
         if self.disabled_or_enabled:
-            if self.simulation_mode:
-                pass
-            else:
-                pass
+            if self.simulation_mode and self.mock_server is None:
+                child = ModbusSlaveContext(hr=ModbusSequentialDataBlock(0, [1] * 30))
+                context = ModbusServerContext(slaves=child, single=True)
+                self.mock_server = ModbusTcpServer(
+                    context=context, address=("127.0.0.1", 15020)
+                )
+                self.log.debug("Starting mock server.")
+                self.mock_server_task = asyncio.create_task(
+                    self.mock_server.serve_forever()
+                )
+                await self.mock_server.serving
+                self.log.debug("Mock server started.")
             if not self.connected:
-                self.component.connect()
-            if self.telemetry_task.done():
+                await self.component.connect()
+            if self.telemetry_task.done() and self.connected:
                 self.telemetry_task = asyncio.create_task(self.telemetry_loop())
         else:
-            if self.simulator is not None:
-                await self.simulator.close()
-                self.simulator = None
-            self.component.disconnect()
             self.telemetry_task.cancel()
+            if self.connected:
+                await self.component.disconnect()
+            if self.mock_server is not None:
+                await self.mock_server.shutdown()
+                self.mock_server_task.cancel()
+                self.mock_server = None
 
     async def configure(self, config):
         """Configure the GIS."""
@@ -85,4 +129,14 @@ class GISCsc(salobj.ConfigurableCsc):
 
     @staticmethod
     def get_config_pkg():
+        """Return the name of the configuration repository."""
         return "ts_config_ocs"
+
+    async def close_tasks(self):
+        """Shutdown mock server and connection to server."""
+        await super().close_tasks()
+        self.telemetry_task.cancel()
+        await self.component.disconnect()
+        if self.mock_server is not None:
+            await self.mock_server.shutdown()
+            self.mock_server = None