+++ tests/test_csc.py
@@ -1,14 +1,13 @@
-import unittest
-import pathlib
 import os
+import pathlib
+import unittest
 
-from lsst.ts import salobj, gis
-
+from lsst.ts import gis, salobj
 
 TEST_CONFIG_DIR = pathlib.Path(__file__).parents[1].joinpath("tests", "data", "config")
 
 
-class GISCscTestCase(unittest.IsolatedAsyncioTestCase, salobj.BaseCscTestCase):
+class GISCscTestCase(salobj.BaseCscTestCase, unittest.IsolatedAsyncioTestCase):
     def setUp(self) -> None:
         os.environ["LSST_SITE"] = "gis"
         return super().setUp()
@@ -17,7 +16,7 @@ class GISCscTestCase(unittest.IsolatedAsyncioTestCase, salobj.BaseCscTestCase):
         self,
         initial_state,
         config_dir=TEST_CONFIG_DIR,
-        simulation_mode=0,
+        simulation_mode=1,
         override="",
     ):
         return gis.GISCsc(
@@ -28,9 +27,13 @@ class GISCscTestCase(unittest.IsolatedAsyncioTestCase, salobj.BaseCscTestCase):
         )
 
     async def test_standard_state_transitions(self):
-        async with self.make_csc(initial_state=salobj.State.STANDBY):
+        async with self.make_csc(
+            initial_state=salobj.State.STANDBY,
+            simulation_mode=1,
+            config_dir=TEST_CONFIG_DIR,
+        ):
             await self.check_standard_state_transitions(
-                enabled_commands=[], override=""
+                enabled_commands=[], override="", timeout=20
             )
 
     async def test_bin_script(self):
@@ -39,6 +42,17 @@ class GISCscTestCase(unittest.IsolatedAsyncioTestCase, salobj.BaseCscTestCase):
             exe_name="run_gis",
             index=None,
         )
+
+    async def test_telemetry(self):
+        async with self.make_csc(
+            initial_state=salobj.State.ENABLED,
+            simulation_mode=1,
+            config_dir=TEST_CONFIG_DIR,
+        ):
+            await self.remote.evt_rawStatus.aget(timeout=20)
+            system_status = await self.remote.evt_systemStatus.aget(timeout=20)
+            assert system_status.index == 28
+            assert system_status.status == 1
 
 
 if __name__ == "__main__":