+++ doc/version-history.rst
@@ -10,6 +10,14 @@ Version History
 .. No new work should be required in order to complete this section.
 .. Below is an example of a version history format.
 
+v0.4.0
+======
+* Swap ts-salobj base case.
+* Fix import path for pymodbus.
+* Add mock_server and simulation mode.
+* Switch to async modbus client.
+* Add sshtunnel functionality
+
 v0.3.0
 ======
 * Use pyproject.toml